Fixed Price Polymorph Error

+++ src/main/Price.java
@@ -52,10 +52,9 @@ class RegularPrice extends Price {
 
     @Override
     double getCharge(int daysRented) {
-        int result = 2;
+        double result = 2;
         if (daysRented > 2)
             result += (daysRented - 2) * 1.5;
-
         return result;
     }
 }
